requirements.txt
streamlit>=1.32.0
Pillow>=10.0.0
pytesseract>=0.3.10
